+++ tests/test_locator.py
@@ -297,7 +297,7 @@ async def test_locator_do_work_transfer_request_fc_exception(config, mocker):
 @pytest.mark.asyncio
 async def test_locator_do_work_transfer_request_fc_no_results(config, mocker):
     """Test that _do_work_transfer_request raises an exception when the LTA DB refuses to create an empty list."""
-    QUARANTINE = {'status': 'quarantined', 'reason': 'File Catalog returned zero files for the TransferRequest'}
+    QUARANTINE = {'status': 'quarantined', 'reason': 'File Catalog returned zero files for the TransferRequest', 'work_priority_timestamp': mocker.ANY}
     logger_mock = mocker.MagicMock()
     p = Locator(config, logger_mock)
     lta_rc_mock = mocker.MagicMock()