+++ CHANGELOG.md
@@ -18,6 +18,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Security
 - in case of vulnerabilities
 
+## [0.5.0] - 2020-03-26
+### Added
+- New work_priority_timestamp to LTA DB
+- New commands: ltacmd {bundle,request} priority reset
+### Changed
+- Modified command to display priority date for bundle and request
+
 ## [0.4.0] - 2020-03-26
 ### Changed
 - Refactored Deleter component to be make it ready for Kubernetes deployment
@@ -232,7 +239,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Project setup scripts
 
-[Unreleased]: https://github.com/WIPACrepo/lta/compare/v0.4.0...HEAD
+[Unreleased]: https://github.com/WIPACrepo/lta/compare/v0.5.0...HEAD
+[0.5.0]: https://github.com/WIPACrepo/lta/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/WIPACrepo/lta/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/WIPACrepo/lta/compare/v0.2.1...v0.3.0
 [0.2.1]: https://github.com/WIPACrepo/lta/compare/v0.2.0...v0.2.1